% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Ensemble Learning for Model Combination}
\usage{
ensemble(
  X.train,
  y.train,
  X.test,
  y.test = NULL,
  model.specs,
  combine.method = "stacking"
)
}
\arguments{
\item{model.specs}{A list of lists, where each inner list specifies the model type and its parameters. For a given model, user must always specify the parameters as NULL if they are not used in the model. Example: list(list(model="linear", alpha = NULL, lambda = NULL,intercept = TRUE, kernel = NULL), list(model="logistic", alpha = NULL, lambda = NULL,intercept = TRUE, kernel = NULL)).}

\item{combine.method}{The method to be used for combining model predictions. Supported methods are "average" for simple averaging of predictions and "stacking" for combine predictions based on their performance using Meta learner which is an SVM.}

\item{X_train}{A matrix of predictor variables for training the models.}

\item{y_train}{A vector of the response variable used for training models.}

\item{X_test}{A matrix of predictor variables for testing the models.}

\item{y_test}{A vector of the response variable used for evaluating the models}
}
\value{
A vector of combined predictions based on the specified combination method.
}
\description{
This function constructs an ensemble of various predictive models specified by the user and combines their predictions using either averaging or SVM-based stacking. It supports multiple types of models, including linear regression, logistic regression, ridge regression, lasso regression, elastic net, and support vector machines.
}
\examples{
{
  # Example using mtcars dataset
  X <- as.matrix(mtcars[, 1:7])
  y <- mtcars$mpg  # Using mpg as the response variable
  specs <- list(list(model="linear"), list(model="ridge", lambda=0.1))
  predictions <- ensemble(X, y, specs, combine.method = "average")

  # Using SVM stacking
  specs <- list(list(model="linear"), list(model="lasso", lambda=0.05))
  predictions <- ensemble(X, y, specs, combine.method = "svm_stacking")
}
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[stats]{glm}}, \code{\link[stats]{svm}}
}
