% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{fit.model}
\alias{fit.model}
\title{Fit Predictive Models with Custom Parameters}
\usage{
fit.model(
  X,
  y,
  model.type,
  alpha = NULL,
  lambda = NULL,
  intercept = TRUE,
  kernel = NULL
)
}
\arguments{
\item{X}{A matrix of predictor variables.}

\item{y}{A vector of the response variable.}

\item{model.type}{A character string specifying the type of model to be fitted. Supported model types include "linear", "logistic", "ridge", "lasso", "elastic", and "svm".}

\item{alpha}{The mixing parameter, it must be between 0 and 1. Default is 0.5.}

\item{lambda}{The regularization parameter for ridge, lasso, and elastic net models. Default is 0.0001.}

\item{intercept}{Logical, indicates whether an intercept should be included in the model. Default is TRUE.}

\item{kernel}{The type of kernel to be used in SVM models. Supported kernels are "linear", "polynomial", "radial", and "sigmoid". Default is "radial".}
}
\value{
The fitted model object, which varies depending on the model type specified. For linear and logistic regression, returns a `glm` object. For ridge, lasso, and elastic net, returns a `glmnet` object. For SVM, returns an `e1071::svm` object.
}
\description{
This function allows fitting of various types of models including linear, logistic, ridge, lasso, elastic net, and SVM. It supports custom configurations for model parameters like alpha, lambda, and kernel.
}
\examples{
{
  # Example using mtcars dataset
  X <- as.matrix(mtcars[, 1:7])
  y <- mtcars$mpg  # Using mpg as the response variable
  model <- fit.model(X, y, model.type = "ridge")

  # Example with logistic regression using iris dataset
  X <- as.matrix(iris[, 1:4])
  y <- ifelse(iris$Species == "setosa", 1, 0)  # Binary outcome
  logistic_model <- fit.model(X, y, model.type = "logistic")
}
}
\seealso{
\code{\link[glmnet]{glmnet}}, \code{\link[e1071]{svm}}
}
